[build-scripts] Use git submodules for all layers
Use git submodules for OpenXT layers, the same way as upstream
layers are configured.

Use a dedicated per-build branch in the build scripts repository in
the local git mirror to set active layers as git submodules.

Use the same dedicated branch to configure BBLAYERS in bblayers.conf
and supply the list of images to be built in build/conf/images.conf

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>